Val_Projects/cpu_stress_validator/requirements.txt
psutil
pyyaml
pytest